+++ src/main/java/fr/inria/corese/gui/editor/button/ValidateShaclButton.java
@@ -58,7 +58,6 @@ public class ValidateShaclButton extends Button {
 
                         // Eval
                         Shacl shacl = new Shacl(coreseGraph, shapeGraph);
-                        shacl.setDataManager(mainFrame.getMyCorese().getDataManager());
 
                         Graph result = null;
                         try {